Make tests on separate file
