+++ src/__tests__/countComments.test.js
@@ -1,0 +1,8 @@
+import { likeslist, mealList, commentList } from '../__mocks__/fetchApi.js';
+import countComments from '../modules/countComments.js';
+
+test('count comments', () => {
+  const comments = commentList();
+  const count = countComments(comments);
+  expect(count).toBe(3);
+});
