+++ src/__tests__/countMeals.test.js
@@ -1,0 +1,9 @@
+import { likeslist, mealList, commentList } from '../__mocks__/fetchApi.js';
+import countMeals from '../modules/countMeals.js';
+
+
+test('count meals on page', () => {
+  const meals = mealList();
+  const count = countMeals(meals);
+  expect(count).toBe(4);
+});
