+++ src/modules/countComments.js
@@ -1,0 +1,5 @@
+const countComments = (comments) => {
+  return comments.length || 0;
+};
+
+export default countComments;
