+++ src/modules/countMeals.js
@@ -1,0 +1,6 @@
+const countMeals = (mealData) => {
+  return mealData.length
+}
+
+
+export default countMeals
